+++ index.js
@@ -1,6 +1,7 @@
 const Container = require('./src/Container');
 const Keyline = require('./src/Keyline');
 const Layout = require('./src/Layout');
+const SpacingTokens = require('./src/SpacingTokens');
 const Spacing = require('./src/Spacing');
 const Typography = require('./src/Typography');
 const RatioBox = require('./src/RatioBox');
@@ -16,6 +17,7 @@ module.exports = {
   Container,
   Keyline,
   Layout,
+  SpacingTokens,
   Spacing,
   Typography,
   RatioBox,