+++ changelog.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2021-08-02
+
+#### Added
+
+- Added spacing tokens plugin to convert spacing tokens from `px` to `rem`
+
+#### Changed
+
+- Typography plugin converts `px` to `rem`
+- Spacing plugin converts `px` to `rem`
+- Spacing tokens added to tailwind config sample
+- Spacing tokens and spacing groups split out in config JSON
+
+
 ## [2.0.2] - 2021-06-17
 
 #### Changed